+++ fishbase/__init__.py
@@ -22,6 +22,7 @@ from .fish_date import *
 from .fish_file import *
 from .fish_logger import *
 from .fish_system import *
+from .fish_project import *
 
-__version__ = '1.0.13'  # type: str
+__version__ = '1.1.1'  # type: str