+++ fishbase/fish_common.py
@@ -16,13 +16,21 @@ import re
 import hashlib
 import hmac
 import os
+import base64
+import string
+import random
 from collections import OrderedDict
+from operator import attrgetter
 import functools
 
 if sys.version > '3':
     import configparser
+    from urllib.parse import parse_qs, urlsplit, urlencode
 else:
     import ConfigParser as configparser
+    from urllib import urlencode
+    from urlparse import parse_qs, urlsplit
+    
 
 # uuid kind const
 udTime = 10001
@@ -42,15 +50,18 @@ charNum = 10022
 # 2018.2.12 #11014 edit by David Yi, 增加返回内容，字典长度,
 # 2018.4.18 #19015 加入 docstring，完善文档说明
 # 2018.5.14 v1.0.11 #19028 逻辑修改，更加严密
-def conf_as_dict(conf_filename):
+# v1.0.15 edit by Hu Jun, #83
+# v1.0.16 edit by Hu Jun, #94
+def conf_as_dict(conf_filename, encoding=None):
     """
     读入 ini 配置文件，返回根据配置文件内容生成的字典类型变量；
 
     :param:
         * conf_filename: (string) 需要读入的 ini 配置文件长文件名
+        * encoding: (string) 文件编码
     :return:
         * flag: (bool) 读取配置文件是否正确，正确返回 True，错误返回 False
-        * d: (dict) 如果读取配置文件正确返回的包含配置文件内容的字典
+        * d: (dict) 如果读取配置文件正确返回的包含配置文件内容的字典，字典内容顺序与配置文件顺序保持一致
         * count: (int) 读取到的配置文件有多少个 key 的数量
 
     举例如下::
@@ -103,14 +114,17 @@ def conf_as_dict(conf_filename):
 
     # 读入 config 文件
     try:
-        cf.read(conf_filename)
+        if sys.version > '3':
+            cf.read(conf_filename, encoding=encoding)
+        else:
+            cf.read(conf_filename)
     except:
         flag = False
         return flag,
 
-    d = dict(cf._sections)
+    d = OrderedDict(cf._sections)
     for k in d:
-        d[k] = dict(cf._defaults, **d[k])
+        d[k] = OrderedDict(cf._defaults, **d[k])
         d[k].pop('__name__', None)
 
     flag = True
@@ -235,14 +249,54 @@ def get_uuid(kind):
 get_time_uuid = functools.partial(get_uuid, udTime)
 
 
-# 功能：判断参数列表是否存在不合法的参数，如果存在None或空字符串或空格字符串，则返回True, 否则返回False
-# 输入参数：source 是参数列表或元组
-# 输出参数：True : 有元素为 None，或空； False：没有元素为 None 或空
 # 2017.2.22 edit by David.Yi, #19007
+# 2018.6.29 v1.0.14 edit by Hu Jun，#62
 def if_any_elements_is_space(source):
-    for i in source:
+    """
+    判断对象中的元素，如果存在None或空字符串或空格字符串，则返回True, 否则返回False, 支持字典、列表和元组
+
+    :param:
+        * source: (list, set, dict) 需要检查的对象
+
+    :return:
+        * result: (bool) 存在None或空字符串或空格字符串返回True， 否则返回False
+
+    举例如下::
+
+        print('--- if_any_elements_is_space demo---')
+        print(if_any_elements_is_space([1, 2, 'test_str']))
+        print(if_any_elements_is_space([0, 2]))
+        print(if_any_elements_is_space([1, 2, None]))
+        print(if_any_elements_is_space((1, [1, 2], 3, '')))
+        print(if_any_elements_is_space({'a': 1, 'b': 0}))
+        print(if_any_elements_is_space({'a': 1, 'b': []}))
+        print('---')
+
+    执行结果::
+
+        --- if_any_elements_is_space demo---
+        False
+        False
+        True
+        True
+        False
+        True
+        ---
+
+    """
+    if isinstance(source, dict):
+        check_list = list(source.values())
+    elif isinstance(source, list) or isinstance(source, tuple):
+        check_list = list(source)
+    else:
+        raise TypeError('source except list, tuple or dict, but got {}'.format(type(source)))
+    
+    for i in check_list:
+        if i is 0:
+            continue
         if not (i and str(i).strip()):
             return True
+    
     return False
 
 
@@ -418,6 +472,7 @@ def if_json_contain(left_json, right_json, op='strict'):
 
 # 2018.3.8 edit by Xiang qinqin
 # 2018.5.15 edit by David Yi, #19030
+# v1.0.15 edit by Hu Jun, #67
 def splice_url_params(dic):
     """
     根据传入的键值对，拼接 url 后面 ? 的参数，比如 ?key1=value1&key2=value2
@@ -445,11 +500,10 @@ def splice_url_params(dic):
     od = OrderedDict(sorted(dic.items()))
 
     url = '?'
-    for key, value in od.items():
-        temp_str = key + '=' + value
-        url = url + temp_str + '&'
-    # 去掉最后一个&字符
-    url = url[:len(url) - 1]
+    temp_str = urlencode(od)
+    
+    url = url + temp_str
+    
     return url
 
 
@@ -496,14 +550,14 @@ def sorted_list_from_dict(p_dict, order=odASC):
 
 
 # v1.0.13 edit by David Yi, edit by Hu Jun，#36
-def check_str(p_str, check_style=charChinese):
+# v1.0.14 edit by Hu Jun #38
+def is_contain_special_char(p_str, check_style=charChinese):
     """
     检查字符串是否含有指定类型字符
     
     :param:
         * p_str: (string) 需要判断的字符串
-        * check_style: (string) 需要判断的字符类型，默认为 charChinese，检查是否含有中文，编码仅支持utf-8，
-        支持 charNum，检查是否含有数字字符串，该参数向后兼容
+        * check_style: (string) 需要判断的字符类型，默认为 charChinese(编码仅支持utf-8),支持 charNum，该参数向后兼容
 
     :return:
         * True 含有指定类型字符
@@ -511,7 +565,7 @@ def check_str(p_str, check_style=charChinese):
 
     举例如下::
         
-        print('--- check_str demo ---')
+        print('--- is_contain_special_char demo ---')
         p_str1 = 'meiyouzhongwen'
         non_chinese_result = check_str(p_str1, check_style=charChinese)
         print(non_chinese_result)
@@ -531,7 +585,7 @@ def check_str(p_str, check_style=charChinese):
 
     执行结果::
         
-        --- check_str demo ---
+        --- is_contain_special_char demo ---
         False
         True
         False
@@ -630,3 +684,321 @@ def hmac_sha256(secret, message):
                           message.encode('utf-8'),
                           digestmod=hashlib.sha256).hexdigest()
     return hashed_str
+
+
+# v1.0.14 edit by Hu Jun, #59
+class Base64:
+    """
+    计算返回文件和字符串的base64编码字符串
+
+    举例如下::
+
+        print('--- Base64 demo ---')
+        print('string base64:', Base64.string('hello world!'))
+        print('file base64:', Base64.file(get_abs_filename_with_sub_path('test_conf', 'test_conf.ini')[1]))
+        print('decode base64:', Base64.decode(b'aGVsbG8gd29ybGQ='))
+        print('---')
+
+    执行结果::
+
+        --- Base64 demo ---
+        string base64: b'aGVsbG8gd29ybGQ='
+        file base64: b'IyEvYmluL2Jhc2gKCmNkIC9yb290L3d3dy9zaW5nbGVfcWEKCm5vaHVwIC9yb290L2FwcC9weXRob24zNjIvYmluL2d1bmljb3JuIC1jIGd1bmljb3JuLmNvbmYgc2luZ2xlX3NlcnZlcjphcHAK'
+        decode base64: b'hello world'
+        ---
+
+    """
+    
+    @staticmethod
+    def string(s):
+        """
+        获取一个字符串的base64值
+
+        :param:
+            * (string) s 需要进行 base64编码 的字符串
+        :return:
+            * (bytes) base64 编码结果
+        """
+        return base64.b64encode(s.encode('utf-8'))
+    
+    @staticmethod
+    def file(filename):
+        """
+        获取一个文件的base64值
+
+        :param:
+            * (string) filename 需要进行 base64编码 文件路径
+        :return:
+            * (bytes) base64 编码结果
+        """
+        with open(filename, 'rb') as f:
+            return base64.b64encode(f.read())
+    
+    @staticmethod
+    def decode(s):
+        """
+        获取base64 解码结果
+
+        :param:
+            * (string) filename 需要进行 base64编码 文件路径
+        :return:
+            * (bytes) base64 解码结果
+        """
+        return base64.b64decode(s)
+
+
+# v1.0.14 edit by Hu Jun, #51
+def get_random_str(length, letters=True, digits=False, punctuation=False):
+    """
+    获得指定长度，不同规则的随机字符串，可以包含数字，字母和标点符号
+    
+    :param:
+        * length: (int) 随机字符串的长度
+        * letters: (bool) 随机字符串是否包含字母，默认包含
+        * digits: (bool) 随机字符串是否包含数字，默认不包含
+        * punctuation: (bool) 随机字符串是否包含特殊标点符号，默认不包含
+
+    :return:
+        * random_str: (string) 指定规则的随机字符串
+
+    举例如下::
+
+        print('--- get_random_str demo---')
+        print(get_random_str(6))
+        print(get_random_str(6, digits=True))
+        print(get_random_str(12, punctuation=True))
+        print(get_random_str(6, letters=False, digits=True))
+        print(get_random_str(12, letters=False, digits=True, punctuation=True))
+        print('---')
+
+    执行结果::
+
+        --- get_random_str demo---
+        nRBDHf
+        jXG5wR
+        )I;rz{ob&Clg
+        427681
+        *"4$0^`2}%9{
+        ---
+
+    """
+    random_source = ''
+    random_source += string.ascii_letters if letters else ''
+    random_source += string.digits if digits else ''
+    random_source += string.punctuation if punctuation else ''
+
+    random_str = ''.join(random.sample(random_source, length))
+    return random_str
+
+
+# v1.0.15 edit by Hu Jun, #77 #63
+def remove_duplicate_elements(items, key=None):
+    """
+    去除序列中的重复元素，使得剩下的元素仍然保持顺序不变，对于不可哈希的对象，需要指定key，说明去重元素
+
+    :param:
+        * items: (list) 需要去重的列表
+        * key: (hook函数) 指定一个函数，用来将序列中的元素转换成可哈希类型
+
+    :return:
+        * result: (generator) 去重后的结果的生成器
+
+    举例如下::
+
+        print('--- remove_duplicate_elements demo---')
+        list_demo = remove_duplicate_elements([1, 5, 2, 1, 9, 1, 5, 10])
+        print(list(list_demo))
+        list2 = [{'x': 1, 'y': 2}, {'x': 1, 'y': 3}, {'x': 1, 'y': 2}, {'x': 2, 'y': 4}]
+        dict_demo1 = remove_duplicate_elements(list2, key=lambda d: (d['x'], d['y']))
+        print(list(dict_demo1))
+        dict_demo2 = remove_duplicate_elements(list2, key=lambda d: d['x'])
+        print(list(dict_demo2))
+        dict_demo3 = remove_duplicate_elements(list2, key=lambda d: d['y'])
+        print(list(dict_demo3))
+        print('---')
+
+    执行结果::
+
+        --- remove_duplicate_elements demo---
+        [1, 5, 2, 9, 10]
+        [{'x': 1, 'y': 2}, {'x': 1, 'y': 3}, {'x': 2, 'y': 4}]
+        [{'x': 1, 'y': 2}, {'x': 2, 'y': 4}]
+        [{'x': 1, 'y': 2}, {'x': 1, 'y': 3}, {'x': 2, 'y': 4}]
+        ---
+
+    """
+    seen = set()
+    for item in items:
+        val = item if key is None else key(item)
+        if val not in seen:
+            yield item
+            seen.add(val)
+
+
+# v1.0.15 edit by Hu Jun, #64
+def sorted_objs_by_attr(objs, key, reverse=False):
+    """
+    对原生不支持比较操作的对象根据属性排序
+
+    :param:
+        * objs: (list) 需要排序的对象列表
+        * key: (string) 需要进行排序的对象属性
+        * reverse: (bool) 排序结果是否进行反转，默认为False，不进行反转
+
+    :return:
+        * result: (list) 排序后的对象列表
+
+    举例如下::
+
+        print('--- sorted_objs_by_attr demo---')
+
+
+        class User(object):
+            def __init__(self, user_id):
+                self.user_id = user_id
+                
+        
+        users = [User(23), User(3), User(99)]
+        result = sorted_objs_by_attr(users, key='user_id')
+        reverse_result = sorted_objs_by_attr(users, key='user_id', reverse=True)
+        print([item.user_id for item in result])
+        print([item.user_id for item in reverse_result])
+        print('---')
+
+    执行结果::
+
+        --- sorted_objs_by_attr demo---
+        [3, 23, 99]
+        [99, 23, 3]
+        ---
+
+    """
+    if len(objs) == 0:
+        return []
+    if not hasattr(objs[0], key):
+        raise AttributeError('{0} object has no attribute {1}'.format(type(objs[0]), key))
+    result = sorted(objs, key=attrgetter(key), reverse=reverse)
+    return result
+
+
+# v1.0.15 edit by Hu Jun, #79
+def get_query_param_from_url(url):
+    """
+    从url中获取query参数字典
+
+    :param:
+        * url: (string) 需要获取参数字典的url
+
+    :return:
+        * query_dict: (dict) query参数的有序字典，字典的值为query值组成的列表
+
+    举例如下::
+
+        print('--- get_query_param_from_url demo---')
+        url = 'http://localhost:8811/mytest?page_number=1&page_size=10&start_time=20180515&end_time=20180712'
+        query_dict = get_query_param_from_url(url)
+        print(query_dict['page_size'])
+        print('---')
+
+    执行结果::
+
+        --- get_query_param_from_url demo---
+        ['10']
+        ---
+
+    """
+    url_obj = urlsplit(url)
+    query_dict = parse_qs(url_obj.query)
+
+    return OrderedDict(query_dict)
+
+
+# v1.1.0 edit by Hu Jun, #74
+def get_group_list_data(data_list, group_number=1, group_size=10):
+    """
+    获取分组列表数据
+
+    :param:
+        * data_list: (list) 需要获取分组的数据列表
+        * group_number: (int) 分组信息，默认为1
+        * group_size: (int) 分组大小，默认为10
+
+    :return:
+        * group_data: (list) 分组数据
+
+    举例如下::
+
+        print('--- get_group_list_data demo---')
+        all_records = [1, 2, 3, 4, 5]
+        print(get_group_list_data(all_records))
+        
+        all_records1 = list(range(100))
+        print(get_group_list_data(all_records1, group_number=5, group_size=15))
+        print(get_group_list_data(all_records1, group_number=7, group_size=15))
+        print('---')
+
+    执行结果::
+
+        --- get_group_list_data demo---
+        [1, 2, 3, 4, 5]
+        [60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74]
+        [90, 91, 92, 93, 94, 95, 96, 97, 98, 99]
+        ---
+
+    """
+    if not isinstance(data_list, list):
+        raise TypeError('data_list should be a list, but we got {}'.format(type(data_list)))
+        
+    if not isinstance(group_number, int) or not isinstance(group_size, int):
+        raise TypeError('group_number and group_size should be int, but we got group_number: {0}, '
+                        'group_size: {1}'.format(type(group_number), type(group_size)))
+    if group_number < 0 or group_size < 0:
+        raise ValueError('group_number and group_size should be positive int, but we got '
+                         'group_number: {0}, group_size: {1}'.format(group_number, group_size))
+
+    start = (group_number - 1) * group_size
+    end = group_number * group_size
+
+    return data_list[start:end]
+
+
+# v1.1.0 edit by Hu Jun, #89
+def get_sub_dict(data_dict, key_list, default_value='default_value'):
+    """
+    从字典中提取子集
+
+    :param:
+        * data_dict: (dict) 需要提取子集的字典
+        * key_list: (list) 需要获取子集的键列表
+        * default_value: (string) 当键不存在时的默认值，默认为 default_value
+
+    :return:
+        * sub_dict: (dict) 子集字典
+
+    举例如下::
+
+        print('--- get_sub_dict demo---')
+        dict1 = {'a': 1, 'b': 2, 'list1': [1,2,3]}
+        list1 = ['a', 'list1', 'no_key']
+        print(get_sub_dict(dict1, list1))
+        print(get_sub_dict(dict1, list1, default_value='new default'))
+        print('---')
+
+    执行结果::
+
+        --- get_sub_dict demo---
+        {'a': 1, 'list1': [1, 2, 3], 'no_key': 'default_value'}
+        {'a': 1, 'list1': [1, 2, 3], 'no_key': 'new default'}
+        ---
+
+    """
+    if not isinstance(data_dict, dict):
+        raise TypeError('data_dict should be dict, but we got {}'.format(type(data_dict)))
+
+    if not isinstance(key_list, list):
+        raise TypeError('key_list should be list, but we got {}'.format(type(key_list)))
+
+    sub_dict = dict()
+    for item in key_list:
+        sub_dict.update({item: data_dict.get(item, default_value)})
+    return sub_dict